update: use `multiprocessing` for playing sound.
`soundcard` is really sensitive to latency. Python's `threading` isn't
parallel, so whenever stuff happens like new input is received or the
window resized, the audio would glitch. Sometimes it would glitch even
if I'm not doing anything. This changes it to run in its own process in
parallel and the playback is much more stable and gapless.