add: ability to accept input over Unix socket.
The app will listen on '/tmp/kokoro-tui' for any incoming UTF-8 text. It
will read until EOF is reached, then create a new audio from the
received text.

I added this to work with the MacOS Shortcuts app, so I can make a
keyboard shortcut to send selected text over the socket to this app to
generate audio. This makes for a much more seemless text-to-speech
experience.