fix: text wrapping not adjusting with resize.
When the text view resized, the text would remain at its previous width.
Sometimes it was longer than the width of the widget and would require
scrolling to view all the text. Sometimes it was shorter and would look
cropped and weird.

This changes the text to always dynamically wrap at the length of the
widget whenever it's resized.