fix: playback choppiness issue.
Audio playback used to have noisy, static-y artifacts whenever the OS
is under load. This commit changes the implementation to queue all
available audio data to the player as soon as it becomes available,
instead of feeding small chunks at a time in a real-time fashion.

This implementation uses a non-public API `_Player._queue` from
`soundcard.coreaudio` so it might be susceptible to breaking. I also
have only tested it on MacOS so I don't know if it'll work on other
platforms.