fix: UI elements not updating under a modal screen.
When some asynchronous action triggers requiring UI updates when a modal
screen is displayed, e.g. `ConfigScreen` or `FilepathInput`, those
updates will not trigger properly because the elements won't be found in
the DOM tree since it searches under the current screen by default,
which is the modal screen and not the app itself. This commit fixes that
issue.